Split keycloak backend out into separate module
