+++ rust/user-info-fetcher/src/backend/mod.rs
@@ -1,0 +1,1 @@
+pub mod keycloak;
